+++ src/common_amirainvest_com/lib/common_amirainvest_com/utils/query_fragments/feed.py
@@ -4,7 +4,6 @@ from sqlalchemy.future import select
 from sqlalchemy.sql import Select
 
 from common_amirainvest_com.schemas import schema
-from common_amirainvest_com.schemas.schema import Bookmarks, Posts, Users, UserSubscriptions
 from common_amirainvest_com.utils.generic_utils import get_past_datetime
 from common_amirainvest_com.utils.sqlalchemy_helpers import DictBundle
 
@@ -12,34 +11,6 @@ from common_amirainvest_com.utils.sqlalchemy_helpers import DictBundle
 PAGE_SIZE = 30
 MAX_HOURS_AGO = 168  # NUMBER OF HOURS TO QUERY POSTGRES : 168H = 1W
 MAX_FEED_SIZE = 200
-
-
-def all_subscriber_posts(subscriber_id: str, hours_ago):
-    return (
-        select(Posts)
-        .join(UserSubscriptions, UserSubscriptions.creator_id == Posts.creator_id)
-        .where(Posts.created_at > get_past_datetime(hours=hours_ago))
-        .where(UserSubscriptions.subscriber_id == subscriber_id)
-        .order_by(Posts.created_at.desc())
-    )
-
-
-def all_creator_posts(creator_id, hours_ago):
-    return (
-        select(Posts)
-        .where(Posts.creator_id == creator_id)
-        .where(Posts.created_at > get_past_datetime(hours=hours_ago))
-        .order_by(Posts.created_at.desc())
-    )
-
-
-def all_user_bookmarked_posts(user_id):
-    return (
-        select(Posts, Bookmarks, Users)
-        .join(Bookmarks.post_id == Posts.id)
-        .join(Bookmarks.user_id == Users.id)
-        .where(Users.id == user_id)
-    )
 
 
 def subscriber_posts(
